+++ MDAnalysisData/PEG_1chain.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
-"""MD simulation of I-FABP with water.
+"""Molecular dynamics trajectory of a single PEG chain in TIP3P water.
 
-https://figshare.com/articles/Molecular_dynamics_trajectory_of_I-FABP_for_testing_and_benchmarking_solvent_dynamics_analysis/7058030
+https://doi.org/10.6084/m9.figshare.7325774
 """
 
 from os.path import dirname, exists, join
@@ -15,8 +15,8 @@ from .base import _fetch_remote, _read_description
 from .base import RemoteFileMetadata
 from .base import Bunch
 
-NAME = "ifabp_water"
-DESCRIPTION = "ifabp_water.rst"
+NAME = "PEG_1chain"
+DESCRIPTION = "PEG_1chain.rst"
 # The original data can be found at the figshare URL.
 # The SHA256 checksum of the zip file changes with every download so we
 # cannot check its checksum. Instead we download individual files.
@@ -24,36 +24,29 @@ DESCRIPTION = "ifabp_water.rst"
 # Bunch that is returned.
 ARCHIVE = {
     'topology': RemoteFileMetadata(
-        filename='ifabp_water.psf',
-        url='https://ndownloader.figshare.com/files/12980639',
-        checksum='ba40714318aabec537015dc550fe5bd5ac1ac0b853f5abdd2f0ae63af9cfcafa',
-    ),
-    'structure': RemoteFileMetadata(
-        filename='ifabp_water_0.pdb',
-        url='https://ndownloader.figshare.com/files/12980636',
-        checksum='8ccf5f75fd85385921c0cb77f00281a93b933fc1261c42fc9492f43983448a72',
+        filename='PEG.prmtop',
+        url='https://ndownloader.figshare.com/files/13532462',
+        checksum='2d7955b9a8cb6e008171e0c5a1c31e3e458246ea3ee7302281eafefafa7cede9',
     ),
     'trajectory':  RemoteFileMetadata(
-        filename='rmsfit_ifabp_water_1.dcd',
-        url='https://ndownloader.figshare.com/files/12980642',
-        checksum='cebb48e58015abc8ff2f5bb7ba3eb7a289047f256351a8252bf1f29f9aaacf0e',
+        filename='PEG_03_prod.nc',
+        url='https://ndownloader.figshare.com/files/13532465',
+        checksum='b978714ec2f93d1cbe99564cb257959f0cb38872359aa745c8eba720a7d85225',
     ),
 }
-
-
 
 logger = logging.getLogger(__name__)
 
 
-def fetch_ifabp_water(data_home=None, download_if_missing=True):
-    """Load the I-FABP with water 0.5 ns equilibrium trajectory
+def fetch_PEG_1chain(data_home=None, download_if_missing=True):
+    """Load the PEG polymer trajectory
 
     Parameters
     ----------
     data_home : optional, default: None
         Specify another download and cache folder for the datasets. By default
         all MDAnalysisData data is stored in '~/MDAnalysis_data' subfolders.
-        This dataset is stored in ``<data_home>/ifabp_water``.
+        This dataset is stored in ``<data_home>/CG_fiber``.
     download_if_missing : optional, default=True
         If ``False``, raise a :exc:`IOError` if the data is not locally available
         instead of trying to download the data from the source site.
@@ -65,13 +58,11 @@ def fetch_ifabp_water(data_home=None, download_if_missing=True):
         Filename of the topology file
     dataset.trajectory : filename
         Filename of the trajectory file
-    dataset.structure : filename
-        Filename of a structure file in PDB format
     dataset.DESCR : string
         Description of the trajectory.
 
 
-    See :ref:`ifabp-water-dataset` for description.
+    See :ref:`PEG_1chain-dataset` for description.
     """
     name = NAME
     data_location = join(get_data_home(data_home=data_home),