+++ MDAnalysisData/datasets.py
@@ -13,6 +13,9 @@ from .adk_transitions import (fetch_adk_transitions_DIMS,
 from .nhaa_equilibrium import fetch_nhaa_equilibrium
 from .ifabp_water import fetch_ifabp_water
 from .vesicles import fetch_vesicle_lib
+from .CG_fiber import fetch_CG_fiber
+from .PEG_1chain import fetch_PEG_1chain
+from .membrane_peptide import fetch_membrane_peptide
 
 __all__ = [
     'get_data_home',
@@ -23,4 +26,7 @@ __all__ = [
     'fetch_ifabp_water',
     'fetch_vesicle_lib'
     'fetch_nhaa_equilibrium',
+    'fetch_CG_fiber',
+    'fetch_PEG_1chain',
+    'fetch_membrane_peptide',
 ]